refactor according to azure python guidelines: https://azure.github.io/azure-sdk/python_design.html#async-support
