+++ .travis.yml
@@ -17,7 +17,7 @@ install:
   - pip install -U pip
   - pip install -r dev_requirements.txt
   - pip install codecov coverage
-  - pip install ./azure-kusto-data[async] ./azure-kusto-ingest
+  - pip install ./azure-kusto-data[aio] ./azure-kusto-ingest
   - pip freeze
 before_script:
   - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then black . --line-length 160 --check; fi
