+++ azure-kusto-data/setup.py
@@ -42,5 +42,5 @@ setup(
     keywords="kusto wrapper client library",
     packages=find_packages(exclude=["azure", "tests"]),
     install_requires=["adal>=1.0.0", "python-dateutil>=2.8.0", "requests>=2.13.0", "msrestazure>=0.4.14"],
-    extras_require={"pandas": ["pandas==0.24.1"], ":python_version<'3.0'": ["azure-nspkg"], "async": ["aiohttp==3.6.2", "asgiref==3.2.3", "aiofile==1.5.2"]},
+    extras_require={"pandas": ["pandas==0.24.1"], ":python_version<'3.0'": ["azure-nspkg"], "aio": ["aiohttp==3.6.2", "asgiref==3.2.3", "aiofile==1.5.2"]},
 )
