+++ azure-kusto-data/tests/aio/case.py
@@ -2,7 +2,7 @@ import unittest
 from asyncio import new_event_loop, set_event_loop
 
 
-class AsyncUnittest(unittest.TestCase):
+class TestCase(unittest.TestCase):
     def setUp(self) -> None:
         self.loop = new_event_loop()
         set_event_loop(self.loop)
